requirements.txt
# Requirements for the project
accelerate>=0.26.0
transformers>=4.30.0
pandas
Pillow
torch
